requirements.txt
# Tomato Disease Agent Chatbot - Main Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
python-dotenv>=1.0.1
tiktoken>=0.7.0

# LangChain + Vector store
langchain>=0.2.12
langchain-core>=0.2.34
langchain-community>=0.2.11
faiss-cpu>=1.8.0.post1

# Hugging Face Transformers
transformers>=4.40.0
torch>=2.3.0
accelerate>=0.30.0
sentence-transformers>=2.2.2

# OpenAI API
openai>=1.30.0

# UI
streamlit>=1.36.0

# Backend API
flask>=2.3.3
flask-cors>=4.0.0
pillow>=10.0.1
tensorflow>=2.13.0
numpy>=1.24.3

# Development
black>=23.7.0
pytest>=7.4.0
